Move state.rs under database module
The types in state are only used for the database
